Fixed Bug: evolutions-list was set to Integer instead String
