--- src/main/java/ch/uzh/ifi/seal/soprafs20/objects/Deck.java
@@ -1,16 +1,0 @@
-package ch.uzh.ifi.seal.soprafs20.objects;
-
-import ch.uzh.ifi.seal.soprafs20.entity.Card;
-
-import java.util.List;
-
-public class Deck {
-    private List<Card> deck;
-
-    public List<Card> getDeck(){ return deck;}
-    public void setDeck(List<Card> deck){this.deck=deck;}
-
-    //peek
-    //remove top
-    //add bottom
-}
