+++ src/main/java/ch/uzh/ifi/seal/soprafs20/rest/dto/GameGetDTO.java
@@ -30,8 +30,6 @@ public class GameGetDTO {
 
     private PlayerDTO turnPlayer;
 
-    private Deck deck;
-
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
 
@@ -67,13 +65,5 @@ public class GameGetDTO {
 
     public void setTurnPlayer(PlayerDTO turnPlayer) {
         this.turnPlayer = turnPlayer;
-    }
-
-    public Deck getDeck() {
-        return deck;
-    }
-
-    public void setDeck(Deck deck) {
-        this.deck = deck;
     }
 }
