+++ src/main/java/ch/uzh/ifi/seal/soprafs20/rest/dto/PlayerDTO.java
@@ -1,5 +1,6 @@
 package ch.uzh.ifi.seal.soprafs20.rest.dto;
 
+import ch.uzh.ifi.seal.soprafs20.entity.Deck;
 import ch.uzh.ifi.seal.soprafs20.entity.Player;
 import ch.uzh.ifi.seal.soprafs20.entity.Statistics;
 import ch.uzh.ifi.seal.soprafs20.rest.mapper.*;
@@ -10,6 +11,7 @@ public class PlayerDTO {
     private UserGetDTO user;
     private Long id;
     private Integer berries;
+    private Deck deck;
 
 
     public UserGetDTO getUser() { return user; }
@@ -24,5 +26,13 @@ public class PlayerDTO {
 
     public void setBerries(Integer berries) {
         this.berries = berries;
+    }
+
+    public Deck getDeck() {
+        return deck;
+    }
+
+    public void setDeck(Deck deck) {
+        this.deck = deck;
     }
 }
