+++ src/main/java/ch/uzh/ifi/seal/soprafs20/rest/mapper/DTOMapper.java
@@ -52,6 +52,7 @@ public interface DTOMapper {
 
     @Mapping(source = "id", target = "id")
     @Mapping(source = "user", target = "user")
+    @Mapping(source = "deck", target = "deck")
     @Mapping(source = "berries", target = "berries")
     PlayerDTO convertEntityToPlayerDTO(Player player);
 
@@ -64,7 +65,6 @@ public interface DTOMapper {
     @Mapping(source = "mode", target = "mode")
     @Mapping(source = "state", target = "state")
     @Mapping(source = "creationTime", target = "creationTime")
-    @Mapping(source = "deck", target = "deck")
     @Mapping(source = "turnPlayer", target = "turnPlayer")
     GameGetDTO convertEntityToGameGetDTO(Game game);
